algo for maze 2 as an arduino code
At the time this code is written the linefollower class did not support backward navigation. So the backward function is written as turning right twice and moving forward. But for task 2 we need to go backwards in order to meet the requirements. Pulasthi can you add a direction parameter to the diveTillJunction function. Need to update the code accordingly after that is implemented. furthermore need to add a functionality to detecting the open gate. the header files does not support that yet